+++ lib/model/setting.dart
@@ -1,6 +1,6 @@
 import 'package:Pilll/model/pill_sheet_type.dart';
 import 'package:Pilll/model/user.dart';
-import 'package:Pilll/repository/user_repository.dart';
+import 'package:Pilll/repository/user.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/foundation.dart';
 import 'package:json_annotation/json_annotation.dart';
